Moved ProjectItemController, changed editor.fxml, added in a couple icons, changed size of Stage.
