+++ .gitignore
@@ -6,7 +6,6 @@
 .bin
 
 .idea/.name
-.idea/misc.xml
 gradle/
 
 # User-specific stuff
