+++ src/main/java/controllers/ProjectItemController.java
@@ -1,4 +1,4 @@
-package org.openjfx;
+package controllers;
 
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
